perf(mtp): graph-capture the batched K=4 verify + batch the GDN conv/WY across sequences
Eliminates the two sized eager gaps in the batched verify step (validator:
~22 ms/step of ungraphed launch gaps + ~5 ms/step of per-seq GDN small
kernels at cap>1):

1. Slot-vector-keyed CUDA graphs for decode_verify_batched_k4. The forward
   span (64-layer loop + final norm + lm_head + argmax) is captured once per
   distinct ssm-pool slot vector and replayed after; embed/KV-ensure/
   metadata/WY-table uploads stay pre-graph into fixed scratch addresses and
   the argmax D2H stays post-graph — the decode_a2 padded_n pattern, keyed
   by the slot VECTOR because the graph bakes every sequence's SSM state
   pointers. Cache bounded (32; overflow runs eager). Kill switch
   ATLAS_NO_MTP_VERIFY_GRAPHS (presence). ATLAS_K4_DIAG forces eager.

2. Cross-sequence batched GDN conv+WY in the GdnStates::Multi arm: first
   call site for the shipped gdn_verify_fused_conv_kn_batched (one launch,
   gridDim.y = n, rollback snapshots inline) plus ONE table-form
   gdn_decode_wy4 at batch_size = n via the existing state_is_table
   pointer tables — replacing the per-seq n x (4 conv + 3 D2D + 1 wy4)
   loop (8n launches/layer -> 2). Tables are staged per GDN layer into a
   fixed model buffer (verify_wy_tables) refreshed pre-graph every step.
   Consecutive-slot and intermediates-contiguity preconditions are checked
   on the actual pointers per layer with a logged per-seq fallback. Kill
   switch ATLAS_NO_VERIFY_GDN_BATCH (presence). The batched conv also
   writes the dead t=K-1 snapshot; the pool allocates num_intermediates=K
   slots and the check requires index K-1 before launching.

Plumbing: decode_verify_multi gains a wy_tables slice param (trait default
+ ssm impl + verify_e call site); table layout constants are SSOT in
crate::layer. Also fixes a pre-existing clippy deny (unnecessary u32 cast
in mtp_head/forward_batch.rs) that broke cargo clippy on the branch.

Left eager by design: Phase-1 host picks, verdict-dependent stash rows,
secondary-stream verdict commits, the batched propose's per-position
syncs, and the argmax readback. Build green (qwen3.6-27b, cuda); GPU
validation (kill-switch A/B, coherence smoke, ENGAGED-log proof of kernel
resolution) is owed by the validator.

+++ crates/spark-model/src/layers/mtp_head/forward_batch.rs
@@ -374,7 +374,7 @@ impl MtpHead {
                 hd,
                 bs as u32,
                 inv_sqrt_d,
-                (nq * hd) as u32,
+                nq * hd,
                 0,
                 stream,
             )?;
